refactor: dsl_to_ir
