+++ crates/polars-plan/src/plans/conversion/dsl_to_ir/utils.rs
@@ -1,0 +1,83 @@
+use std::path::PathBuf;
+
+use super::*;
+
+#[derive(Default)]
+pub(super) struct SourcesToFileInfo {
+    inner: PlHashMap<Arc<[PathBuf]>, FileInfo>,
+}
+
+impl SourcesToFileInfo {
+    pub(super) fn insert(&mut self, source: &ScanSources, info: &FileInfo) {
+        match source {
+            ScanSources::Paths(paths) => {
+                self.inner.insert(paths.clone(), info.clone());
+            },
+            ScanSources::Files(_) | ScanSources::Buffers(_) => {
+                // don't save the files or buffers
+            },
+        }
+    }
+
+    pub(super) fn get(&self, source: &ScanSources) -> Option<&FileInfo> {
+        match source {
+            ScanSources::Paths(paths) => self.inner.get(paths),
+            _ => None,
+        }
+    }
+}
+
+pub(super) struct DslConversionContext<'a> {
+    pub(super) expr_arena: &'a mut Arena<AExpr>,
+    pub(super) lp_arena: &'a mut Arena<IR>,
+    pub(super) conversion_optimizer: ConversionOptimizer,
+    pub(super) opt_flags: &'a mut OptFlags,
+    pub(super) nodes_scratch: &'a mut UnitVec<Node>,
+    pub(super) cache_file_info: SourcesToFileInfo,
+    pub(super) pushdown_maintain_errors: bool,
+    pub(super) verbose: bool,
+}
+
+pub(super) fn expand_expressions(
+    input: Node,
+    exprs: Vec<Expr>,
+    lp_arena: &Arena<IR>,
+    expr_arena: &mut Arena<AExpr>,
+    opt_flags: &mut OptFlags,
+) -> PolarsResult<Vec<ExprIR>> {
+    let schema = lp_arena.get(input).schema(lp_arena);
+    let exprs = rewrite_projections(exprs, &schema, &[], opt_flags)?;
+    to_expr_irs(exprs, expr_arena, &schema)
+}
+
+pub(super) fn empty_df() -> IR {
+    IR::DataFrameScan {
+        df: Arc::new(Default::default()),
+        schema: Arc::new(Default::default()),
+        output_schema: None,
+    }
+}
+
+pub(super) fn validate_expression(
+    node: Node,
+    expr_arena: &Arena<AExpr>,
+    input_schema: &Schema,
+    operation_name: &str,
+) -> PolarsResult<()> {
+    let iter = aexpr_to_leaf_names_iter(node, expr_arena);
+    validate_columns_in_input(iter, input_schema, operation_name)
+}
+
+pub(super) fn validate_expressions<N: Into<Node>, I: IntoIterator<Item = N>>(
+    nodes: I,
+    expr_arena: &Arena<AExpr>,
+    input_schema: &Schema,
+    operation_name: &str,
+) -> PolarsResult<()> {
+    let nodes = nodes.into_iter();
+
+    for node in nodes {
+        validate_expression(node.into(), expr_arena, input_schema, operation_name)?
+    }
+    Ok(())
+}
